third_party/WebKit/LayoutTests/http/tests/inspector-enabled/console-promise-reject-and-handle.html

<html>
<head>
<script src="../inspector/inspector-test.js"></script>
<script src="../inspector/console-test.js"></script>
<script>
if (window.testRunner) {
    testRunner.waitUntilDone();
    testRunner.dumpAsText();
}

function setupPromise1()
{
    var p = Promise.reject(new Error('e'));
    window.onunhandledrejection = function(evt) {
        console.log("onunhandledrejection1");
        evt.preventDefault();
        setTimeout(function() {
            p.then(function(){}, function(){});
        }, 10);
    };
    window.onrejectionhandled = function(evt) {
        console.log("onrejectionhandled1");
        setTimeout(setupPromise2, 0);
    };
}

function setupPromise2()
{
    var p = Promise.reject(new Error('e'));
    window.onunhandledrejection = function(evt) {
        console.log("onunhandledrejection2");
        setTimeout(function() {
            p.then(function(){}, function(){});
        }, 10);
    };
    window.onrejectionhandled = function(evt) {
        console.log("onrejectionhandled2");
        setTimeout(startTest, 0);
    };
    function startTest()
    {
        if (window.testRunner)
            testRunner.showWebInspector();
        runTest();
    }
}

function test() {
  ConsoleTestRunner.expandConsoleMessages();
  TestRunner.addResult('----console messages start----');
  ConsoleTestRunner.dumpConsoleMessages();
  TestRunner.addResult('----console messages end----');
  TestRunner.completeTest();
}
</script>
</head>
<body onload="setupPromise1()">
<p>
Tests that evt.preventDefault() in window.onunhandledrejection suppresses console output.
</p>
</body>
</html>
